requirements.txt
# Observe Insurance ARIA — LangGraph Multi-Agent Voice AI
# pip install -r requirements.txt

# ── LangGraph + LangChain core ────────────────────────────────────────────────
langgraph==0.2.28
langchain==0.3.7
langchain-core==0.3.15

# ── LLM (OpenRouter or Groq — both OpenAI-compatible) ─────────────────────────
openai==1.51.0

# ── Voice STT ─────────────────────────────────────────────────────────────────
deepgram-sdk==3.7.3

# ── Voice TTS ─────────────────────────────────────────────────────────────────
elevenlabs==1.9.0

# ── Telephony — Twilio (real phone number, works on Windows) ──────────────────
twilio==9.3.3
websockets==13.1
flask==3.0.3                 # Webhook server Twilio calls back to
flask-sock==0.7.0            # WebSocket support for real-time audio streaming

# ── Audio processing ──────────────────────────────────────────────────────────
numpy==1.26.4
audioop-lts==0.2.1           # Python 3.13 compat for audio encoding (mulaw/pcm)

# ── HTTP + async ──────────────────────────────────────────────────────────────
aiohttp==3.10.10
httpx==0.27.2

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic>=2.10.6
nest-asyncio==1.6.0          # Run async graph inside Flask sync context

# ── Google Sheets setup script only ──────────────────────────────────────────
gspread==6.1.4
google-auth==2.35.0
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.12
